requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy>=1.21.0
matplotlib>=3.5.0
PyYAML>=6.0
tqdm>=4.65.0
pillow>=9.0.0
scikit-learn>=1.0.0
jupyter>=1.0.0
pytest>=7.0.0